added findUser Api and general fixes

+++ src/routes/todo.ts
@@ -1,23 +1,26 @@
-import express from "express";
+import express,{ Request, Response } from "express";
 import { TodoCRUD } from '../crud/TodoCRUD';
 import { TodoRepository } from '../repositories/TodoRepository';
 import { Todo } from '../entities/mongo/todoSchema';
 import { TodoEntity } from '../entities/TodoEntity';
 import { MongoDataStorage } from '../dataStorages/MongoDataStorage';
+import { authMiddleware } from "../utils/authMiddleware";
 
 const routes = express.Router()
 
 const REPOSITORY = new TodoRepository(new MongoDataStorage<TodoEntity>(Todo));
+routes.use(authMiddleware);
 
-routes.get('/', async (req, res) => {
-    const todos = await new TodoCRUD(REPOSITORY).read();
+routes.get('/', async (req: Request, res: Response) => {
+    const todos = await new TodoCRUD(REPOSITORY).read({userId: req.userId});
 
     res.status(todos.statusCode).json(todos.data);
 });
 
 routes.post('/', async (req, res) => {
     const { text } = req.body;
-    const insertNewTodo = await new TodoCRUD(REPOSITORY).create({ text });
+    const userId = req.userId
+    const insertNewTodo = await new TodoCRUD(REPOSITORY).create({text, userId});
 
     res.status(insertNewTodo.statusCode).json(insertNewTodo.data);
 });
@@ -34,7 +37,8 @@ routes.put('/:id', async (req, res) => {
 });
 
 routes.delete('/deleteAll', async (req, res) => {
-    const deletedTodos = await new TodoCRUD(REPOSITORY).deleteAll()
+    const userId = req.userId
+    const deletedTodos = await new TodoCRUD(REPOSITORY).deleteAll({userId})
 
     res.status(deletedTodos.statusCode).json(deletedTodos.data)
 })
